Additional image member information
Return the created_at and updated_at values for image member.

Related to bp glance-api-v2-image-sharing

Change-Id: I50f514c8680d1299d41c107c993e7ee2b0d7a185

Image Member Domain Model

This commit is the first layer of image member domain model.
The can_share feature from v1 of glance api will not be
ported to glance v2 api.

Related to bp glance-api-v2-image-sharing

Change-Id: Ibfb790d438efb0f5b29b4786c02f1155962150c5

Adding database layer for image members domain model

Adding the database layer image members domain model.
For a given image, a user can get a list of image members,
delete image member, and get a specific image.

Related to bp glance-api-v2-image-sharing

Change-Id: I71a0c0d3c828cba231a75d0b7888bad932de41e5

Adding image members in glance v2 api

Adding image members to glance v2 api. A user can create image member,
list image members and delete image member for a given image.
Also adds the necessary authorization.

Related to bp glance-api-v2-image-sharing

Change-Id: Ifc0607d693f7b0218ebd7fc4824a64e9eba995b3

Add status column to image_members

Related to bp glance-api-v2-image-sharing

Change-Id: Id8896f9dcb717ab8a466829b3b1cd9fde4b67a36

Allowing member to set status of image membership

Only the owner of an image can add members to the image.
By default the status of the image member is requested.
The member can change the status of the image membership
to accepted or rejected. This is done to prevent spamming
the user's image list. A member cannot see the other members
of the image.

Related to bp glance-api-v2-image-sharing

Change-Id: I0d0deba4b0df52b2f8d105b779fb7de746229d3a

Filter images by status

This patch allows the user to filter images by status.
Whenever a user does an image list, the following are displayed:
(1) public images
(2) images owned by the user
(3) images whom the user is a member of and has set status to accepted
The user can change the status in (3) and also choose to display
the images the user has been requested to be a member of,
rejected to be a member of, or all images irrespective of the status
set by user

Related to bp glance-api-v2-image-sharing

Change-Id: I052c0947f9ad17ef1b14af0ad0f51b886afd0d7a

Filter by owned_by in conjunction with member status

This patch allows the user to filter images by the owner of the image
in conjunction with the member status.
Whenever a user does an image list, the following are displayed:
(1) public images
(2) images owned by the user
(3) images whom the user is a member of and has set status to accepted
The user can change the status in (3) and also choose to display
the images the user has been requested to be a member of,
rejected to be a member of, or all images irrespective of the status
set by user
(3b) owned_by tenant_id allows the user to filter by images whom the
user is a member of and has is owned by a specific owner(tenant_id)

DocImpact

Related to bp glance-api-v2-image-sharing

Change-Id: I752f4a153b91d9723ead6ddc8ce84862e9514d61